modified views file
